requirements.txt
roboticstoolbox-python
numpy==1.26.4
scipy
python-math
matplotlib==3.5.3
pyqt5
pyqt6
tk